1) cors updated 2) updated delete image controller to delete file from disk 3) removed slash from multer config to remove path issue

+++ backend/controllers/imageController.js
@@ -1,4 +1,6 @@
-const Image = require("../models/imageModel")
+const path = require('path');
+const Image = require("../models/imageModel");
+const fs = require('fs/promises');
 
 
 exports.addImage = async (req, res) => {
@@ -9,12 +11,12 @@ exports.addImage = async (req, res) => {
             return res.status(400).json({ message: "Please upload a file" });
         }
 
-        const baseURL = `${req.protocol}://${req.get('host')}`;
+        const baseURL = `https://${req.get('host')}`;
 
         const imageDocs = files.map((file) => ({
             originalname: file.originalname,
             filename: file.filename,
-            path: `${baseURL}/${file.path}`,
+            path: `${baseURL}/uploads/${file.filename.replace(/\\/g, '/')}`, 
             size: file.size,
             caption
         }));
@@ -33,11 +35,9 @@ exports.getAllImages = async (req, res) => {
         let images = await Image.find()
         let totalImages = images.length
         return res.status(200).json({ message: "images fetched successfully", success: true, totalImages, images })
-
     } catch (error) {
         console.log(error.message)
         return res.status(500).json({ message: "images fetch failed", success: false })
-
     }
 }
 exports.getOneImage = async (req, res) => {
@@ -50,20 +50,26 @@ exports.getOneImage = async (req, res) => {
         return res.status(200).json({ image })
     } catch (error) {
         console.log(error.message)
-        return res.status(500).json({ message: "Error while fetching image with id: " + id })
+        return res.status(500).json({ message: `Error while fetching image with id: ${id}`, success: false })
     }
 }
 
 exports.deleteImage = async (req, res) => {
     try {
         let { id } = req.params
-        let deletedImage = await Image.findByIdAndDelete(id)
-        if (!deletedImage) {
-            return res.status(404).json({ message: "Image not found", success: false })
+        let file = await Image.findById(id)
+        if (!file) {
+            return res.status(404).json({ message: "Image not found", success: false });
         }
-        return res.json({ message: "Image deleted successfully", sucees: true, deletedImage })
+
+        let filelocation = path.resolve(__dirname, "..", 'uploads', file.filename);
+
+        await Image.findByIdAndDelete(id);
+        await fs.unlink(filelocation);
+
+        return res.status(200).json({ message: "Image deleted successfully", success: true });
     } catch (error) {
-        console.log(error.message)
-        return res.status(500).json({ message: "Image deleting failed", success: false })
+        console.log(error.message);
+        return res.status(500).json({ message: "Image deleting failed", success: false });
     }
-}
+};
